+++ src/pages/Home/ui/FashionContent/ConditionButtonGroup/ConditionButtonGroup.style.ts
