+++ src/pages/Home/ui/FashionContent/FashionContent.style.ts
