+++ src/pages/Home/ui/FashionContent/Headline/Headline.style.ts
