+++ src/pages/Home/ui/WeatherInformation/WeatherCard/WeatherCard.style.ts
