+++ src/pages/Home/ui/WeatherInformation/WeatherCard/WeatherCard.tsx
